refactor(core.db): remove dead relationships queries
playResultsWithCharts had no consumers: B30 uses playResultsBestWithCharts,
which joins through PlayResultBestRepository + per-item chart lookups. Drop
the query, the PlayResultWithChart DTO, and the RelationshipsDao that then
only held minimumPlayResultPotentialFields - that query moves to a new
PlayResultBestDao, its only consumer.
