+++ core/src/main/kotlin/xyz/sevive/arcaeaoffline/core/database/entities/Relationships.kt
@@ -4,11 +4,6 @@ import androidx.room.ColumnInfo
 import xyz.sevive.arcaeaoffline.core.constants.ArcaeaRatingClass
 import kotlin.uuid.Uuid
 
-data class PlayResultWithChart(
-    val playResult: PlayResult,
-    val chart: Chart? = null,
-)
-
 data class PlayResultBestWithChart(
     val playResultBest: PlayResultCalculated,
     val chart: Chart? = null,
